// link2rail/projects/ekp/src/app/empty-wagon/masterdata/models/latest-customer-order-date-time-request-view.ts
/**
 * Empty wagon internal API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {DemandLocationView} from "../../common/models/demand-location-view";

export interface LatestCustomerOrderDateTimeRequestView {
    /**
     * Demand date.
     */
    deliveryDateTime: Date;
    demandLocation: DemandLocationView;
}