// link2rail/projects/ekp/src/app/empty-wagon/masterdata/models/masterdata-date-for-commercial-location-request-view.ts
/**
 * Empty wagon internal API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {CommercialLocationView} from "../../common/models/commercial-location-view";

export interface MasterdataDateForCommercialLocationRequestView {
    /**
     * date.
     */
    date: Date;
    commercialLocation?: CommercialLocationView;
}